fixed bug in test

+++ src/DOC_GPU/DOCGPU.cpp
@@ -8,13 +8,13 @@
 #include "DOCGPU.h"
 #include "DOCGPU_Kernels.h"
 
-
 # define CUDA_CALL ( x) do { if (( x) != cudaSuccess ) { \
 	printf (" Error at % s :% d\ n" , __FILE__ , __LINE__ ) ;\
 return EXIT_FAILURE ;}} while (0)
 # define CURAND_CALL ( x) do { if (( x) != CURAND_STATUS_SUCCESS ) { \
 printf (" Error at % s :% d\ n" , __FILE__ , __LINE__ ) ;\
 return EXIT_FAILURE ;}} while (0)
+
 
 DOCGPU::DOCGPU(std::vector<std::vector<float>*>* input, float alpha, float beta, float width) {
 	this->data = input;
@@ -59,9 +59,9 @@ std::pair<std::vector<std::vector<float>*>*, std::vector<bool>*> DOCGPU::findClu
 	auto beta = this->beta;
 	auto width = this->width;
 	
-	float d = data->at(0)->size();
-	float r = log2(2*d)/log2(1/(2*beta));
-	float m = pow((2/alpha),2) * log(4);
+	unsigned int d = data->at(0)->size();
+	unsigned int r = log2(2*d)/log2(1/(2*beta));
+	unsigned int m = pow((2/alpha),2) * log(4);
 	
 	unsigned int number_of_ps = 2.0/alpha;
 	unsigned int number_of_samples = number_of_ps*m;
@@ -269,6 +269,8 @@ std::vector<std::pair<std::vector<std::vector<float>*>*, std::vector<bool>*>> DO
 
 
 bool DOCGPU::generateRandomSubSets(DataReader* dataReader){
+	//https://stackoverflow.com/questions/18501081/generating-random-number-within-cuda-kernel-in-a-varying-range
+	// TODO:
 
 	size_t size= 100;
 	curandGenerator_t gen ;
@@ -290,6 +292,7 @@ bool DOCGPU::generateRandomSubSets(DataReader* dataReader){
 
 	// Copy device memory to host
 	cudaMemcpy ( hostData , randomNumberArray_d , size * sizeof ( unsigned int ) ,cudaMemcpyDeviceToHost );
+
 
 	// Cleanup
