+++ src/DOC_GPU/DOCGPU_Kernels.h
@@ -33,19 +33,19 @@ float* scoreHost(unsigned int* Cluster_size,
 
 
 void findDimmensionsKernel(unsigned int dimGrid, unsigned int dimBlock, float* Xs_d, float* ps_d, bool* res_d,
-						   unsigned int* Dsum_out, int point_dim, int no_of_samples, int no_in_sample, int no_of_ps,
-						   float m, float width);
+						   unsigned int* Dsum_out, unsigned int point_dim, unsigned int no_of_samples, unsigned int no_in_sample, unsigned int no_of_ps,
+						   unsigned int m, float width);
 
 void pointsContainedKernel(unsigned int dimGrid, unsigned int dimBlock,
 						   float* data, float* centroids, bool* dims, bool* output, unsigned int* Csum_out,
-									  float width, int point_dim, int no_data, int no_dims, int m);
+									  float width, unsigned int point_dim, unsigned int no_data, unsigned int no_dims, unsigned int m);
 
 void scoreKernel(unsigned int dimGrid, unsigned int dimBlock, unsigned int* Cluster_size, unsigned int* Dim_count, float* score_output,
-					  int len, float alpha, float beta, unsigned int num_points);
+					  unsigned int len, float alpha, float beta, unsigned int num_points);
 
 void createIndicesKernel(unsigned int dimGrid, unsigned int dimBlock, unsigned int* index, unsigned int length);
 
 void argMaxKernel(unsigned int dimGrid, unsigned int dimBlock, unsigned int sharedMemorySize,
-				  float* scores, unsigned int* scores_index, int input_size);
+				  float* scores, unsigned int* scores_index, unsigned int input_size);
 
 #endif
